requirements.txt
ai-data-science-team==0.0.0.9016
streamlit==1.45.0
sqlalchemy==2.0.40
pandas==2.2.3
asyncio==3.4.3
langchain==0.3.25
langchain-openai==0.3.16
langgraph==0.4.1
openai==1.77.0
IPython==8.36.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
langchain_core==0.3.58